# By Shane Parker (231) and others
# Via Shane Parker (9) and others
* 'master' of github.com:shiozaki/BAGEL: (257 commits)
  reordering loops in gamma_forest_prod_asd in order to reduce the number of tranpsposes and GammaForest::search calls
  using the faster apply function for the perturbation
  removing some more allocations inside of the block functions. there's not many left that can be removed.
  making some specialized kronecker functions
  symmetrize function is the slowest part of hamiltonian building
  add fci and BFGS timers to casbfgs.
  avoiding some more allocations and optimizing in block_ops_2 functions
  using the raw pointer rather than Tensor element access for integrals
  avoiding a few allocations in kronecker product
  adding in the option of a more direct kronecker product
  fixing some asserts, using RASBlockVectors directly in more places
  just for good measure
  minor
  switch to using blockinfo rather than RASCivecViews
  optimizing Sparse_IJ a bit
  avoiding building unnecessary RASCivecView objects
  making dedicated functions for the pure hamiltonian parts. also splitting the form_sigma file apart.
  some assert fixes
  a new cleaner and much faster sigma_ab function in RAS
  changes the resolve_S_abb function so it now acts on RASBlockVectors rather than CivecViews
  ...

Conflicts:
	src/Makefile.am

+++ src/fci/fci.cc
@@ -305,7 +305,7 @@ void FCI::compute() {
         double* source_array = errvec[ist]->data();
         double* denom_array = denom_->data();
         const double en = energies[ist];
-        for (int i = 0; i != size; ++i) {
+        for (size_t i = 0; i != size; ++i) {
           target_array[i] = source_array[i] / min(en - denom_array[i], -0.1);
         }
         cc_->data(ist)->normalize();
